lock down basic versions for all plugins utilized in Jenkins development
Locking down every commonly used Maven plugin utilized in Jenkins to avoid
warnings when using Maven 3.x. Since locking down plugin versions helps to
ensure a repeatable build, it is advantageous to do so at the top level of
the project hierarchy so that all sub-modules can take advantage.

This really should not have any meaningful effect on the system because:

 1) a sub project already codes the plugin version, and that version will
    be utilized instead of the default one here, or
 2) no version is coded and so the latest version will be scanned for and
    picked up; since we've locked down the plugin versions to what is
    the latest available, this amounts to a net effect

We should consider running a CI job which runs:

   mvn versions:display-plugin-updates

so that we can get a report on which plugins are available for updating.

The only thing I expect to happen with this is that Maven 3.x users will
see less nagging and that all users will see a marginal performance boost
because scanning for plugin updates will not be performed.