Make it easier to use a different JVM target
Whilst Jenkins defaults will now be java11 some libraries may want to
stay on java8.

Doing so makes backports of any issues much easier (hopefully not
needed!) and the library can take advantage of all the updated plugins
with their fixes.

Whilst this is less than optimal for java8 as it still required a java9+
JVM to be used to compile (as that was when the `release` flag was
introduced it is much better than either pinning the parent to 1.76 (the
last supporting java8), or having to overwrite all the properties to
remove the maven.compiler.release and testRelease ones as well as
overwriting the enforcer configurations (albeit this is simplier with
combine.children)