Use Palantir Java Format 2.73.0 (#702)
https://github.com/palantir/palantir-java-format/releases/tag/2.73.0
shows it was released last week.

The git client plugin use of `spotless` needs the Palantir Java Format
upgrade to 2.73.0 in order to run `spotless` with Java 25.  The Palantir
Java Format upgrade does not seem to be required for Jenkins core to run
`spotless` with Java 25.

Testing done:

Confirmed that `mvn spotless:apply` succeeds with Java 25 EA build 28
on Jenkins core with or without this change.

Confirmed that `mvn spotless:apply` makes no formatting changes in
Jenkins core with this change.